+++ src/main/java/janggi/service/JanggiService.java
@@ -8,6 +8,7 @@ import janggi.infra.transaction.TransactionExecutor;
 import janggi.model.Janggi;
 import janggi.model.Team;
 import janggi.model.board.Board;
+import janggi.model.board.BoardType;
 import janggi.model.board.PlayingBoard;
 import janggi.model.palace.Palaces;
 import janggi.model.piece.Piece;
@@ -20,7 +21,6 @@ import janggi.model.turn.playing.ChoTurn;
 import janggi.model.turn.playing.HanTurn;
 import janggi.service.dto.GameDetailResponse;
 import janggi.service.dto.GameOptionResponse;
-import janggi.view.mapping.BoardType;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
