+++ src/test/java/janggi/dao/dao/JdbcPieceDaoTest.java
@@ -5,11 +5,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import janggi.dao.DatabaseTest;
 import janggi.dao.piece.PieceEntity;
 import janggi.model.board.Board;
+import janggi.model.board.BoardType;
 import janggi.model.palace.factory.DefaultPalaceFactory;
 import janggi.model.position.absolute.Column;
 import janggi.model.position.absolute.Position;
 import janggi.model.position.absolute.Row;
-import janggi.view.mapping.BoardType;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
