+++ src/test/java/janggi/service/JanggiServiceTest.java
@@ -7,6 +7,7 @@ import janggi.dao.game.GameDao;
 import janggi.dao.piece.PieceDao;
 import janggi.model.Janggi;
 import janggi.model.Team;
+import janggi.model.board.BoardType;
 import janggi.model.board.PlayingBoard;
 import janggi.model.palace.Palaces;
 import janggi.model.palace.factory.DefaultPalaceFactory;
@@ -19,7 +20,6 @@ import janggi.model.position.absolute.Position;
 import janggi.model.position.absolute.Row;
 import janggi.service.dto.GameDetailResponse;
 import janggi.service.dto.GameOptionResponse;
-import janggi.view.mapping.BoardType;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
